split the concern of maintaining the wallet from maintaining the C bindings
